# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: QES Platform API
  description: |
    Modular eIDAS QES Digital-Signing Platform API
    
    This API provides qualified electronic signature services compliant with eIDAS regulation
    and ETSI standards. Support for multiple EU countries with unified interface.
    
    ## Features
    - XAdES-B, XAdES-T, XAdES-LTA signatures
    - PAdES-B, PAdES-T, PAdES-LTA signatures  
    - Multi-tenant SaaS architecture
    - Country-specific QES providers (Freja eID, D-Trust, FNMT)
    - RFC 3161 timestamping
    - Long-term validation (LTV)
    
    ## Authentication
    All API endpoints require Bearer token authentication. Obtain tokens through
    the `/auth/login` endpoint or via OAuth2 flows.
    
    ## Rate Limiting
    API requests are subject to per-tenant rate limiting:
    - Free tier: 100 requests/hour, 20 signatures/day
    - Professional: 1000 requests/hour, 500 signatures/day
    - Enterprise: 10000 requests/hour, 10000 signatures/day
    
    Rate limit headers are included in responses:
    - `X-RateLimit-Limit`: Request limit for current window
    - `X-RateLimit-Remaining`: Remaining requests in current window
    - `X-RateLimit-Reset`: Unix timestamp when window resets
    
  version: 1.0.0
  contact:
    name: QES Platform Support
    email: support@qes-platform.com
    url: https://docs.qes-platform.com
  license:
    name: Commercial License
    url: https://qes-platform.com/license
servers:
...
    description: Production server
...
security:
  - BearerAuth: []
  - TenantHeader: []
...
paths:
...
  /sign:
    post:
      tags: [Signatures]
      summary: Create digital signature
      description: |
        Sign a document with qualified electronic signature.
        Supports XAdES and PAdES formats with B, T, and LTA levels.
      operationId: signDocument
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/SigningRequest'
            encoding:
              document:
                contentType: application/pdf, application/xml, text/xml
      responses:
        '200':
          description: Document signed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SigningResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /signatures/{signatureId}:
    get:
      tags: [Signatures]
      summary: Get signature details
      operationId: getSignature
      parameters:
        - name: signatureId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Signature details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignatureInfo'

    delete:
      tags: [Signatures]
      summary: Revoke signature
      description: Revoke a previously created signature
      operationId: revokeSignature
      parameters:
        - name: signatureId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Signature revoked successfully

  # Verification endpoints
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from authentication

    TenantHeader:
      type: apiKey
      in: header
      name: X-Tenant-ID
      description: Tenant ID for multi-tenant access
  schemas:
...
    UserInfo:
      type: object
      properties:
        user_id:
          type: string
        given_name:
          type: string
        family_name:
          type: string
        email:
          type: string
          format: email
        country_code:
          type: string
          minLength: 2
          maxLength: 2
        authentication_level:
          type: string
          description: eIDAS assurance level
    Certificate:
      type: object
      properties:
        certificate_id:
          type: string
        subject_dn:
          type: string
          description: Certificate subject distinguished name
        issuer_dn:
          type: string
          description: Certificate issuer distinguished name
        serial_number:
          type: string
        valid_from:
          type: string
          format: date-time
        valid_to:
          type: string
          format: date-time
        key_usage:
          type: array
          items:
            type: string
        certificate_policies:
          type: array
          items:
            type: string
        is_qualified:
          type: boolean
          description: Whether certificate is qualified for QES
    SigningRequest:
      type: object
      properties:
        document:
          type: string
          format: binary
          description: Document to sign
        document_name:
          type: string
          description: Document filename
        signature_format:
          type: string
          enum: [XAdES-B, XAdES-T, XAdES-LTA, PAdES-B, PAdES-T, PAdES-LTA]
          description: Desired signature format and level
        certificate_id:
          type: string
          description: Certificate to use for signing
        timestamp_url:
          type: string
          format: uri
          description: Custom timestamp authority URL
        signature_policy:
          type: string
          description: Signature policy identifier
        metadata:
          type: object
          additionalProperties: true
          description: Additional metadata for signature
      required: [document, document_name, signature_format]
    SigningResponse:
      type: object
      properties:
        signature_id:
          type: string
          format: uuid
        signed_document:
          type: string
          format: binary
          description: Signed document (base64 encoded)
        signature_format:
          type: string
        timestamp:
          type: string
          format: date-time
        certificate_used:
          $ref: '#/components/schemas/Certificate'
        validation_info:
          type: object
          description: Validation information for LTV
...
    SignatureInfo:
      type: object
      properties:
        signature_id:
          type: string
          format: uuid
        document_name:
          type: string
        signature_format:
          type: string
        created_at:
          type: string
          format: date-time
        signed_by:
          $ref: '#/components/schemas/UserInfo'
        certificate:
          $ref: '#/components/schemas/Certificate'
        validation_status:
          type: string
          enum: [valid, invalid, unknown]
        revocation_status:
          type: string
          enum: [good, revoked, unknown]
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
        trace_id:
          type: string
          description: Request trace ID for debugging
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    RateLimited:
      description: Rate limit exceeded
      headers:
        X-RateLimit-Limit:
          schema:
            type: integer
          description: Request limit
        X-RateLimit-Remaining:
          schema:
            type: integer
          description: Remaining requests
        X-RateLimit-Reset:
          schema:
            type: integer
          description: Reset timestamp
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retry
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...